Implement a class named Building
Constructor attributes:
sqft (Number)
The attribute is stored in an “underscore” attribute version.

A getter and setter is implemented for the attribute.

This class is considered as an abstract class. Any class that extends
from it is made sure to have a method named evacuationWarningMessage.

If a class that extends from it does not have this method, an error with
the message 'Class extending Building must override evacuationWarningMessage'
is thrown.