Build a Teacher interface
firstName(string) and lastName(string). These attributes are only
modifiable when a Teacher is first initialized.

fullTimeEmployee(boolean). This attribute is always defined.

yearsOfExperience(number). This attribute is optional.

location(string). This attribute is always defined.

Any attribute can also be added to the Object like contract(boolean)
without specifying the name of the attribute.