Create functions specific to an employee.
isDirector:
It accepts employee as an argument and is used as a type predicate
and if the employee is a director.

executeWork:
It accepts employee as an argument. If the employee is a Director,
it calls workDirectorTasks and if the employee is a Teacher,
it calls workTeacherTasks.

The Subjects type allows a variable to have the value of either
Math or History.

teachClass:
This function takes todayClass as an argument.
It returns the string Teaching Math if todayClass is Math.
It returns the string Teaching History if todayClass is History.